Inline tests on event emitter as not shared anymore
